pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "zcode"
version = "0.1.0"
description = "A minimal CLI agent harness for learning and experimentation."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic",
    "python-dotenv",
]

[project.scripts]
zcode = "zcode.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
